Add files via upload
The file "run_RUSBoost.m" et al.,  is the Matlab code to run the esemble learning algorithm to predict greenwashing. To run the code file, two additional Matlab files are required: (1) the file "data_reader.m" for reading the data, and (2) the file "evaluate.m" for evaluating model performance.
We also provide training code for other algorithms, including Randomforest, LSTM, SVM, et al.
The file "tune_RUSBoost.m" is the Matlab code to replicate the hyper-parameter tuning for our RUSBoost model. The number of learners/trees is tuned by using the traditional grid search approach. Specifically, the parameter space is manually specified as [100, 200, 300, 400, 500, 600, 700, 800, 900, 1000, 1500, 2000, 2500, 3000]. For each hyper-parameter, the model is trained using the training period 2008-2018, and then evaluated in terms of AUC in the validating year 2019-2022.